+++ src/Components/Header.h
@@ -9,7 +9,7 @@
 
 class Header : public juce::Component {
 public:
-    const int ICON_SIZE = 24;
+    const float ICON_SIZE = 24.0f;
 
     typedef juce::AudioProcessorValueTreeState::SliderAttachment SliderAttachment;
     typedef juce::AudioProcessorValueTreeState::ComboBoxAttachment ComboBoxAttachment;
